Library Service with JPA. Issue with BookCopies, gets  no authorId in tbl_book ... but not sure why it's trying to get the authorId from tbl_book

+++ lms/src/main/java/com/ss/training/spring/lms/dao/AuthorDAO.java
@@ -35,7 +35,7 @@ public class AuthorDAO extends BaseDAO<Author>{
 		List<Author> authors = new ArrayList<>();
 		while(rs.next()){
 			Author author = new Author();
-			author.setAuthorId(rs.getInt("authorId"));
+			author.setAuthorId(rs.getLong("authorId"));
 			author.setAuthorName(rs.getString("authorName"));
 			authors.add(author);
 		}
